Adding Shield-Advanced files to support upcoming TDS blog
Adding:

- drt-removal Lambda Node.js function. This allow DRT permissions
to be revoked from S3 and Shield resources, once a DDoS incident
is confirmed as closed.
- shield-engagement Lambda Node.js function. This automatically
launches a support ticket, routes a message to the DRT SNS topic and
grants DRT for access to Shield and S3 resources, to aid them with
analysis and diag tasks during a DDoS incident.
- add-protections.sh Bash script which automatically adds new tagged
AWS resources as shield protected resources.

Full info/instructions in the individual Readme's.